Remove BaseCollection class
