--- src/Sulu/Bundle/MediaBundle/Resources/config/doctrine/BaseCollection.orm.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd"
-                  xmlns:gedmo="http://gediminasm.org/schemas/orm/doctrine-extensions-mapping">
-
-    <mapped-superclass name="Sulu\Bundle\MediaBundle\Entity\BaseCollection">
-        <id name="id" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-        <field name="style" type="string" column="style" length="255" nullable="true"/>
-        <field name="lft" type="integer" column="lft">
-            <gedmo:tree-left/>
-        </field>
-        <field name="rgt" type="integer" column="rgt">
-            <gedmo:tree-right/>
-        </field>
-        <field name="depth" type="integer" column="depth">
-            <gedmo:tree-level/>
-        </field>
-
-        <field name="key" type="string" column="collection_key" length="191" unique="true" nullable="true"/>
-
-        <many-to-one field="type" target-entity="Sulu\Bundle\MediaBundle\Entity\CollectionType" inversed-by="collections">
-            <join-columns>
-                <join-column name="idCollectionTypes" referenced-column-name="id" nullable="false"/>
-            </join-columns>
-        </many-to-one>
-
-        <gedmo:tree type="nested"/>
-    </mapped-superclass>
-</doctrine-mapping>
